delete individual meal pages, create AddFoodPopUpComponent
