+++ src/components/AddFoodPopUpComponent/index.js
@@ -1,0 +1,2 @@
+import AddFoodPopUpComponent from "./AddFoodPopUpComponent";
+export default AddFoodPopUpComponent;
