--- src/pages/AddBreakfastPage/index.js
@@ -1,2 +1,0 @@
-import AddBreakfastPage from "./AddBreakfastPage";
-export default AddBreakfastPage;
