--- src/pages/AddDinnerPage/index.js
@@ -1,2 +1,0 @@
-import AddDinnerPage from './AddDinnerPage';
-export default AddDinnerPage;
