--- src/pages/AddLunchPage/AddLunchPage.jsx
@@ -1,31 +1,0 @@
-import React, { Component } from "react";
-import { connect } from "react-redux";
-import FoodSearchComponent from "../../components/FoodSearchComponent";
-
-class AddLunchPage extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return (
-      <div>
-        <h1>Lunch</h1>
-        <FoodSearchComponent meal_type_id={"2"} />
-      </div>
-    );
-  }
-}
-
-const mapStateToProps = store => {
-  return {};
-};
-
-const mapDispatchToProps = dispatch => {
-  return {};
-};
-
-AddLunchPage = connect(mapStateToProps, mapDispatchToProps)(AddLunchPage);
-
-export default AddLunchPage;
