--- src/pages/AddLunchPage/index.js
@@ -1,2 +1,0 @@
-import AddLunchPage from './AddLunchPage';
-export default AddLunchPage;
