--- src/pages/AddSnackPage/index.js
@@ -1,2 +1,0 @@
-import AddSnackPage from "./AddSnackPage";
-export default AddSnackPage;
